Move string functionality to pkg/strutil
