+++ pkg/strutil/strconv.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package utility
+package strutil
 
 import (
 	"fmt"
