+++ promql/parse.go
@@ -21,8 +21,8 @@ import (
 	"time"
 
 	clientmodel "github.com/prometheus/client_golang/model"
+	"github.com/prometheus/prometheus/pkg/strutil"
 	"github.com/prometheus/prometheus/storage/metric"
-	"github.com/prometheus/prometheus/utility"
 )
 
 type parser struct {
@@ -982,7 +982,7 @@ func (p *parser) checkType(node Node) (typ ExprType) {
 }
 
 func parseDuration(ds string) (time.Duration, error) {
-	dur, err := utility.StringToDuration(ds)
+	dur, err := strutil.StringToDuration(ds)
 	if err != nil {
 		return 0, err
 	}
