+++ rules/recording.go
@@ -21,8 +21,8 @@ import (
 
 	clientmodel "github.com/prometheus/client_golang/model"
 
+	"github.com/prometheus/prometheus/pkg/strutil"
 	"github.com/prometheus/prometheus/promql"
-	"github.com/prometheus/prometheus/utility"
 )
 
 // A RecordingRule records its vector expression into new timeseries.
@@ -90,9 +90,9 @@ func (rule RecordingRule) HTMLSnippet(pathPrefix string) template.HTML {
 	ruleExpr := rule.vector.String()
 	return template.HTML(fmt.Sprintf(
 		`<a href="%s">%s</a>%s = <a href="%s">%s</a>`,
-		pathPrefix+strings.TrimLeft(utility.GraphLinkForExpression(rule.name), "/"),
+		pathPrefix+strings.TrimLeft(strutil.GraphLinkForExpression(rule.name), "/"),
 		rule.name,
 		rule.labels,
-		pathPrefix+strings.TrimLeft(utility.GraphLinkForExpression(ruleExpr), "/"),
+		pathPrefix+strings.TrimLeft(strutil.GraphLinkForExpression(ruleExpr), "/"),
 		ruleExpr))
 }
